% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIDS.R
\docType{data}
\name{AIDS}
\alias{AIDS}
\title{AIDS}
\format{A data frame with 2 columns and 14 rows. The columns are:
\itemize{
\item{TIME: }{Time measured in multiples of 3 months after January 1983.}
\item{DEATHS: }{Number of deaths in Australia due to AIDS.}
}}
\usage{
data(AIDS)
}
\description{
The data set was reported by Whyte et al. (1987). The data content the number of deaths due to AIDS in Australia per 3-month period from January 1983 to June 1986.
}
\details{

}
\examples{
data(AIDS)
}
\references{
Whyte, B., J. Gold, A. Dobson, and D. Cooper (1987). Epidemiology of acquired im- munodeficiency syndrome in Australia. The Medical Journal of Australia 146(2), 65–69.
}
\keyword{datasets}
