% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{articles}
\alias{articles}
\title{articles}
\format{A data frame with 6 columns and 915 rows. The columns are:
\itemize{
\item{fem: }{Gender; 0 = male; 1 = female.}
\item{ment: }{Number of articles published by a scientist’s mentor.}
\item{phd: }{Prestige of the graduate program.}
\item{mar: }{Marital status; 0 = No; 1 = Yes.}
\item{kid5: }{Number of young children.}
\item{art: }{Number of articles.}
}}
\usage{
data(articles)
}
\description{
The data set for a study of the number of articles published by the scientist.
}
\details{
This study is to examine how factors such as gender (fem), marital status (mar), number of young children (kid5), prestige of the graduate program (phd), and number of articles published by a scientist’s mentor (ment) affect the number of articles (art) published by the scientist.
}
\examples{
data(articles)
}
\references{
Long, J. S. (1997). Regression Models for Categorical and Limited Dependent Variables. Thousand Oaks: Sage Publications.
}
\keyword{datasets}
