% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFW.R
\name{BayesLM.nprior}
\alias{BayesLM.nprior}
\title{Bayes Linear Model}
\usage{
BayesLM.nprior(lmfit, B)
}
\arguments{
\item{lmfit}{a linear model object returned by \code{"lm"}}

\item{B}{number of samples to generate}
}
\value{
A list with the elements
\itemize{
\item{post.dist}{data frame containing samples from the posterior of the parameters.}
\item{summary.stat}{summary statistics for the posterior samples}
}
}
\description{
Fits a Bayes Linear Model using direct sampling
}
\examples{
cars.lm <- lm(dist ~ speed, data=cars)
set.seed(1)
cars.blm <- BayesLM.nprior(cars.lm,10000)

}

