% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr.R
\name{bri.adapt.prior}
\alias{bri.adapt.prior}
\title{Adaptive prior}
\usage{
bri.adapt.prior(x, degree = 3, nknot = 5, theta.prec = 0.01,
  type = c("indpt", "spde", "stat"))
}
\arguments{
\item{type}{}
}
\value{
spde object
}
\description{
Adaptive prior
}

