% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr.R
\name{bri.band.plot}
\alias{bri.band.plot}
\title{Plot credible band for nonlinear functionTitle}
\usage{
bri.band.plot(result, name = NULL, alpha = 0.05, ind = NULL,
  xlab = NULL, ylab = NULL, main = NULL, sub = NULL, xlim = NULL,
  ylim = NULL, type = c("random", "fitted", "linear"), hpd = FALSE)
}
\arguments{
\item{hpd}{}
}
\description{
Plot credible band for nonlinear functionTitle
}

