% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XW.R
\name{bri.basehaz.plot}
\alias{bri.basehaz.plot}
\title{Plot the baseline hazard functions for survival models using INLA}
\usage{
bri.basehaz.plot(inla.obj, plot = TRUE, cex.lab = 1.25, cex.axis = 1.25,
  ...)
}
\arguments{
\item{inla.obj}{an object of class "inla".}

\item{plot}{if plot = TRUE, the plot of the baseline hazard functions will be generated.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.axis}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{...}{}
}
\value{
A list of evaluated time points (time), and the values of corresponding base line function (basehaz).
}
\description{
Plot the baseline hazard functions for survival models using INLA
}
\details{
The function supports Weibull model, Expontenial model and Cox proportional hazards model for right censored data.
}
