% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XW.R
\name{bri.beta.resid}
\alias{bri.beta.resid}
\title{Baysian residuals for beta regression using INLA}
\usage{
bri.beta.resid(inla.obj, covariate = NULL, m = 1000, pmedian = FALSE,
  plot = FALSE, smooth = FALSE, xlab = NULL, cex.lab = 1.25,
  cex.axis = 1.25, ylab = "Bayesian residual", ylim = NULL, main = "",
  ...)
}
\arguments{
\item{inla.obj}{an object of class "inla".}

\item{covariate}{the covariate of interest for the residual plot.}

\item{m}{the size of posterior samples to be generated from the posterior distribution.}

\item{pmedian}{if pmedian = TRUE, posterior medians of Bayesian residuals will be calculated.}

\item{plot}{if plot = TRUE, a Bayesian residual plot will be generated}

\item{smooth}{if smooth = TRUE, a nonparametric smooth curve will be added.}

\item{xlab}{a title for the x axis.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.axis}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{ylab}{a title for the y axis.}

\item{main}{an overall title for the plot.}

\item{...}{}
}
\value{
a list consisting of residuals (resid), the covariate used for plot (covariate), the response variable (y)
}
\description{
Baysian residuals for beta regression using INLA
}
\details{
compute Bayesian residuals, and generate a index plot or a plot of the residual versus the predictor for Bayesian beta regression using INLA.
}
\author{
Xiaofeng Wang, \email{wangx6@ccf.org}
}
