% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XW.R
\name{bri.csresid.plot}
\alias{bri.csresid.plot}
\title{Cox-Snell residual plot}
\usage{
bri.csresid.plot(resid.obj, lwd = 2, xlab = "Cox-Snell Residual",
  ylab = "Estimated Cumulative Hazard Rates", main = "", cex.lab = 1.25,
  cex.axis = 1.25, ...)
}
\arguments{
\item{resid.obj}{object from bri.surv.resid function}

\item{lwd}{the line width, a positive number, defaulting to 2.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{main}{an overall title for the plot.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.axis}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{...}{}
}
\description{
Cox-Snell residual plot
}
