% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XW.R
\name{bri.density}
\alias{bri.density}
\title{Density estimation using INLA}
\usage{
bri.density(x, m = 101, from, to, cut = 0.1, diagonal = 0.001,
  constr = T, ...)
}
\arguments{
\item{x}{the data from which the estimate is to be computed.}

\item{m}{the number of equally spaced points at which the density is to be estimated.}

\item{from, to}{the left and right-most points of the grid at which the density is to be estimated.}

\item{cut}{by default, the values of from and to are cut * diff(range(x)) beyond the extremes of the data.}

\item{diagonal}{An extra constant added to the diagonal of the precision matrix in INLA.}

\item{constr}{A boolean variable indicating whater to set a sum to 0 constraint on the term.}

\item{...}{}
}
\value{
a list containing the following components: x - the n coordinates of the points where the density is estimated; y - the estimated density values; y.lower - the estimated 2.5 percentile of density; y.lower - the estimated 97.5 percentile of density.
}
\description{
Density estimation using INLA
}
