% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JJF.R
\name{bri.density.summary}
\alias{bri.density.summary}
\title{Compute a summary from a density}
\usage{
bri.density.summary(dens)
}
\arguments{
\item{dens}{a density}
}
\value{
numerical summary
}
\description{
Compute a summary from a density
}
