% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{bri.excursions.ggplot}
\alias{bri.excursions.ggplot}
\title{Excursions}
\usage{
bri.excursions.ggplot(res.exc, xlab = NULL, ylab = NULL, main = NULL)
}
\arguments{
\item{res.exc}{}

\item{xlab}{}

\item{ylab}{}

\item{main}{}
}
\description{
Excursions
}
