% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JJF.R
\name{bri.gpr}
\alias{bri.gpr}
\title{Gaussian Process Regression in 1D}
\usage{
bri.gpr(x, y, pcprior, nbasis = 25, degree = 2, alpha = 2, xout = x,
  sigma0 = sd(y), rho0 = 0.25 * (max(x) - min(x)))
}
\arguments{
\item{x}{the predictor vector}

\item{y}{the response vector}

\item{pcprior}{limites for the penalised complexity prior (optional). If specified should be a vector 
of the form c(r,s) where P(range < r = 0.05) and P(SD(y) > s = 0.05)}

\item{nbasis}{- number of basis functions for the spline (default is 25)}

\item{degree}{- degree for splines (default is 2) - allowable possibilities are 0, 1 or 2.}

\item{alpha}{- controls shape of the GP kernel (default is 2) - 0 < alpha <=2 is possible}

\item{xout}{- grid on which posterior will be calculated (default is x)}

\item{sigma0}{- prior mean for the signal SD (default is SD(y))}

\item{rho0}{- prior mean for the range}
}
\value{
list consisting of xout, the posterior mean, the lower 95% credibility band, 
the upper 95% credibility band and the INLA object containing the fit
}
\description{
Gaussian Process Regression in 1D
}
\author{
Julian Faraway, \email{jjf23@bath.ac.uk}
}
