% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JJF.R
\name{bri.hyper.sd}
\alias{bri.hyper.sd}
\title{Convert precision to SD}
\usage{
bri.hyper.sd(prec, internal = FALSE)
}
\arguments{
\item{prec}{a precision density}

\item{internal}{logical indicating whether this is an internal representation}
}
\value{
an SD density
}
\description{
Convert precision to SD
}

