% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bri.hyperpar.plot}
\alias{bri.hyperpar.plot}
\title{Plot the hyperparameter posteriors}
\usage{
bri.hyperpar.plot(r, together = TRUE)
}
\arguments{
\item{r}{an INLA model object}

\item{together}{TRUE if densities to be plotted on a single panel}
}
\value{
data frame containing the densities
}
\description{
Plot the hyperparameter posteriors
}

