% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JJF.R
\name{bri.hyperpar.summary}
\alias{bri.hyperpar.summary}
\title{Convert precisions to SD in INLA hyperparameter summary}
\usage{
bri.hyperpar.summary(r)
}
\arguments{
\item{inla}{model object}
}
\value{
summary of hyperparameters on SD scale (where appropriate)
}
\description{
Convert precisions to SD in INLA hyperparameter summary
}
\author{
Julian Faraway, \email{jjf23@bath.ac.uk}
}
