% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFW.R
\name{bri.lmresid.plot}
\alias{bri.lmresid.plot}
\title{linear regression model Baysian residual plot}
\usage{
bri.lmresid.plot(inla.obj, covariate = NULL, m = 1000, pmedian = FALSE,
  smooth = FALSE, xlab = NULL, cex.lab = 1.25, cex.axis = 1.25,
  ylab = "Bayesian residual", main = "", ...)
}
\arguments{
\item{inla.obj}{}

\item{covariate}{}

\item{m}{}

\item{pmedian}{}

\item{smooth}{}

\item{xlab}{}

\item{cex.lab}{}

\item{cex.axis}{}

\item{ylab}{}

\item{main}{}

\item{...}{}
}
\description{
linear regression model Baysian residual plot
}
