% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFW.R
\name{bri.mresid.plot}
\alias{bri.mresid.plot}
\title{Martingale Residual plot}
\usage{
bri.mresid.plot(resid.obj, covariate = NULL, smooth = FALSE, xlab = NULL,
  ylab = "Martingale residual", main = "", cex.lab = 1.25,
  cex.axis = 1.25, ...)
}
\arguments{
\item{resid.obj}{}

\item{covariate}{}

\item{smooth}{}

\item{xlab}{}

\item{ylab}{}

\item{main}{}

\item{cex.lab}{}

\item{cex.axis}{}

\item{...}{}
}
\description{
Martingale Residual plot
}
