% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JJF.R
\name{bri.nonstat}
\alias{bri.nonstat}
\title{Non-stationary smoothing for Gaussian Process Regression in 1D}
\usage{
bri.nonstat(x, y, nbasis = 25, sbasis = 5, degree = 2, alpha = 2,
  xout = x, sigma0 = sd(y), rho0 = 0.25 * (max(x) - min(x)))
}
\arguments{
\item{x}{the predictor vector}

\item{y}{the response vector}

\item{nbasis}{- number of basis functions for the spline (default is 25)}

\item{sbasis}{- number of basis functions for the smoothing of sigma and rho}

\item{degree}{- degree for splines (default is 2) - allowable possibilities are 0, 1 or 2.}

\item{alpha}{- controls shape of the GP kernel (default is 2) - 0 < alpha <=2 is possible}

\item{xout}{- grid on which posterior will be calculated (default is x)}
}
\value{
list consisting of xout, the posterior mean, the lower 95% credibility band, 
the upper 95% credibility band and the INLA object containing the fit
}
\description{
Non-stationary smoothing for Gaussian Process Regression in 1D
}
\author{
Julian Faraway, \email{jjf23@bath.ac.uk}
}
