% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bri.random.plot}
\alias{bri.random.plot}
\title{Plot the posterior densities of the random effects}
\usage{
bri.random.plot(r)
}
\arguments{
\item{r}{inla model object}
}
\value{
a data frame with the densities and group labels
}
\description{
Plot the posterior densities of the random effects
}

