% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JJF.R
\name{bri.smoothband}
\alias{bri.smoothband}
\title{Smoothness bands for Gaussian Process Regression}
\usage{
bri.smoothband(x, y, nbasis = 25, degree = 2, alpha = 2, xout = x,
  sigma0 = sd(y), rho0 = 0.25 * (max(x) - min(x)))
}
\arguments{
\item{x}{the predictor vector}

\item{y}{the response vector}

\item{nbasis}{- number of basis functions for the spline (default is 25)}

\item{degree}{- degree for splines (default is 2) - allowable possibilities are 0, 1 or 2.}

\item{alpha}{- controls shape of the GP kernel (default is 2) - 0 < alpha <=2 is possible}

\item{xout}{- grid on which posterior will be calculated (default is x)}

\item{sigma0}{- prior mean for the signal SD (default is SD(y))}

\item{rho0}{- prior mean for the range}
}
\value{
list consisting of xout, the posterior mean, the smoother 95\% credibility band, 
the rougher 95\% credibility band
}
\description{
Smoothness bands for Gaussian Process Regression
}
\author{
Julian Faraway, \email{jjf23@bath.ac.uk}
}
