% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFW.R
\name{bri.surv.resid}
\alias{bri.surv.resid}
\title{this function is to obtain the different residuals for inla survival models
the function supports Weibull model, Expontenial model and Cox proportional hazards model for right censored data.
Three types of residuals could be obtained: Cox-Snell residual; Martingale residuals; Deviance residuals.}
\usage{
bri.surv.resid(inla.obj, time, event)
}
\arguments{
\item{inla.obj}{}

\item{time}{}

\item{event}{}
}
\description{
this function is to obtain the different residuals for inla survival models
the function supports Weibull model, Expontenial model and Cox proportional hazards model for right censored data.
Three types of residuals could be obtained: Cox-Snell residual; Martingale residuals; Deviance residuals.
}
