% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bronch.R
\docType{data}
\name{bronch}
\alias{bronch}
\title{bronch}
\format{A data frame with 4 columns and 1246 rows. The columns are:
\itemize{
\item{cbr: }{Chronic bronchitis reaction; 0 = No; 1 =Yes.}
\item{dust: }{Dust concentration at work place in mg/m^3.}
\item{smoking: }{whether the worker smokes or not; 0 = No; 1 =Yes.}
\item{expo: }{Duration of exposure.}
}}
\usage{
data(bronch)
}
\description{
The bronchitis data set that has been discussed by Carroll et al., 2006.
}
\details{
The bronchitis study was to assess the health hazard of specific harmful substances in a dust-laden mechanical engineering plant in Munich. The outcome of the study is the indicator that the worker has bronchitis.
}
\examples{
data(bronch)
}
\references{
Carroll, R. J., D. Ruppert, L. A. Stefanski, and C. Crainiceanu (2006). Measurement Error in Nonlinear Models: A Modern Perspective (2nd ed.). New York: Chapman & Hall/CRC Press.
}
\keyword{datasets}
