% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calder.R
\docType{data}
\name{calder}
\alias{calder}
\title{Maximum flows on the River Calder}
\format{A data frame with 3 columns and 41 rows. The columns are:
\itemize{
\item{date}{ISO date of maximum flow that water year}
\item{WaterYear}{water year which runs from Oct 1st of the previous year to Sep 30th of the current year}
\item{Flow}{maximum water flow in cubic meters per second}}}
\source{

}
\usage{
data(calder)
}
\description{
Maximum annual flows on the River Calder in Cumbria
}
\details{
Measurement takes places at Calder Hall, grid reference NY035044
}
\examples{
data(calder)

}
\references{
Data obtained from \url{http://nrfa.ceh.ac.uk/data/station/info/74006}
}
\keyword{datasets}

