% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{default.knots}
\alias{default.knots}
\title{Creat the vector of default knots
An internal function for spline.mixed()}
\usage{
default.knots(x, num.knots)
}
\description{
Creat the vector of default knots
An internal function for spline.mixed()
}
