% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil.R
\docType{data}
\name{fossil}
\alias{fossil}
\title{Strontium content of fossils}
\format{A data frame with 2 columns and 106 rows. The columns are:
\itemize{
\item{age}{ in millions of years.}
\item{sr}{ strontium ratio.}}}
\source{
Bralower, T., P. Fullagar, C. Paull, G. Dwyer, and R. Leckie (1997). Mid-cretaceous strontium-
isotope stratigraphy of deep-sea sections. Geological Society of America Bulletin 109(11), 1421–1442.
}
\usage{
data(fossil)
}
\description{
Strontium content of fossils from the mid-Cretaceous period.
}
\details{
The ratio of strontium isotopes found in fossil shells in the mid-Cretaceous
period from about 90 to 125 million years ago.
}
\examples{
data(fossil)

}
\references{
Chaudhuri, P. and J. S. Marron (1999). Sizer for exploration of structures in curves.
Journal of the American Statistical Association 94 (447), 807–823.
}
\keyword{datasets}
