% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framingham.R
\docType{data}
\name{framingham}
\alias{framingham}
\title{framingham}
\format{A data frame with 7 columns and 1615 rows. The columns are:
\itemize{
\item{AGE: }{Age at exam 2.}
\item{SBP21: }{First systolic blood pressure at exam 2.}
\item{SBP22: }{Second systolic blood pressure at exam 2.}
\item{SBP31: }{First systolic blood pressure at exam 3.}
\item{SBP32: }{Second systolic blood pressure at exam 3.}
\item{SMOKE: }{present smoking at exam 1; 0 = No; 1 = Yes.}
\item{FIRSTCHD: }{Indicator of first evidence of CHD occurring at exam 3 through 6; 0 = No; 1 = Yes.}
}}
\usage{
data(framingham)
}
\description{
The data set is from the Framingham Heart Study (Carroll et al., 2006).
}
\details{
The Framingham study consists of a series of exams taken two years apart. There are 1615 males aged from 31 to 65 in this dataset, with the outcome indicating the occurrence of coronary heart disease (CHD) within an eight-year period following Exam 3. There were 128 total cases of CHD.
}
\examples{
data(framingham)
}
\references{
Carroll, R. J., D. Ruppert, L. A. Stefanski, and C. Crainiceanu (2006). Measurement Error in Nonlinear Models: A Modern Perspective (2nd ed.). New York: Chapman & Hall/CRC Press.
}
\keyword{datasets}
