% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frencheconomy.R
\docType{data}
\name{frencheconomy}
\alias{frencheconomy}
\title{frencheconomy}
\format{A data frame with 5 columns and 18 rows. The columns are:
\itemize{
\item{YEAR: }{Year.}
\item{IMPORT: }{Import.}
\item{DOPROD: }{Domestic production.}
\item{STOCK: }{Stock formation.}
\item{CONSUM: }{Domestic consumption.}
}}
\usage{
data(frencheconomy)
}
\description{
Data for import activity in the French economy. The data have been analyzed in Chatterjee and Hadi (2015).
}
\details{
The dependent variable is import (IMPORT), domestic production (DOPROD), stock formation (STOCK), and domestic consumption (CONSUM). All variables are measured in billions of French francs for the years 1949 to 1966.
}
\examples{
data(frencheconomy)
}
\references{
Chatterjee, S. and A. S. Hadi (2015). Regression Analysis by Example (5th ed.). New York: John Wiley & Sons.
}
\keyword{datasets}
