% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inla.density.summary}
\alias{inla.density.summary}
\title{Compute a summary from a density}
\usage{
inla.density.summary(dens)
}
\arguments{
\item{dens}{a density}
}
\value{
numerical summary
}
\description{
Compute a summary from a density
}

