% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inla.hyperpar.summary}
\alias{inla.hyperpar.summary}
\title{Convert precisions to SD in INLA hyperparameter summary}
\usage{
inla.hyperpar.summary(r)
}
\arguments{
\item{inla}{model object}
}
\value{
summary of hyperparameters on SD scale (where appropriate)
}
\description{
Convert precisions to SD in INLA hyperparameter summary
}

