% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyphosis.R
\docType{data}
\name{kyphosis}
\alias{kyphosis}
\title{Children who have had Corrective Spinal Surgery}
\format{A data frame with 4 columns and 81 rows. The columns are:
\itemize{
\item{Age}{in months.}
\item{NumVert}{The number of vertebrae involved.}
\item{StartVert}{the number of the first (topmost) vertebra operated on.}
\item{Kyphosis}{takes the value 0 if absent or 1 if present after surgery}}}
\source{
Also found in the \code{rpart} package
}
\usage{
data(kyphosis)
}
\description{
Kyphosis is excessive outward curvature of the spine, causing hunching of the back.
}
\examples{
data(kyphosis)
}
\references{
John M. Chambers and Trevor J. Hastie eds. (1992)
Statistical Models in S, Wadsworth and Brooks/Cole, Pacific Grove, CA.
}
\keyword{datasets}

