% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larynx.R
\docType{data}
\name{larynx}
\alias{larynx}
\title{larynx}
\format{The ohio data frame has 2148 rows and 4 columns
\itemize{
\item{resp}{an indicator of wheeze status (1=yes, 0=no).}
\item{id}{a numeric vector for subject id.}
\item{age}{a numeric vector of age, 0 is 9 years old}
\item{smoke}{an indicator of maternal smoking at the first year of the study}
}}
\usage{
data(larynx)
}
\description{
The dataset is
}
\examples{
data(larynx)

}
\references{
here
}
\keyword{datasets}

