% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{map.munich}
\alias{map.munich}
\title{Draw Munich map}
\usage{
map.munich(results, ...)
}
\arguments{
\item{results}{the vector of values to be shown on the map}
}
\value{
A map of Munich with color scale
}
\description{
Draw Munich map
}
