% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzunemploy.R
\docType{data}
\name{nzunemploy}
\alias{nzunemploy}
\title{nzunemploy}
\format{A data frame with 4 columns and 102 rows. The columns are:
\itemize{
\item{quarter: }{Quarters from March 1986 to June 2011.}
\item{adult: }{Unemployment rate for adult.}
\item{youth: }{Unemployment rate for adult.}
\item{policy: }{Type of minimum wage law; "Different" or "Equal".}
}}
\usage{
data(nzunemploy)
}
\description{
The New Zealand unemployment data include the quarterly unemployment rates for both youth (15 – 19 years old) and adult (greater than 19 years old) from March 1986 to June 2011.
}
\details{
Since June 2008, the New Zealand government has abolished the act of the differential youth minimum wage.
}
\examples{
data(nzunemploy)
}
\references{

}
\keyword{datasets}
