% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\docType{data}
\name{reading}
\alias{reading}
\title{Reading scores in school children}
\format{A data frame with 3 columns and 267 rows. The columns are:
\itemize{
\item{id}{the ID of the child}
\item{agegrp}{The age of the child (rounded).}
\item{piat}{The reading score of the child (peabody individual achievment test.}
}}
\source{
Applied Longitudinal Data Analysis: Modeling Change and Event Occurrence by Judith D. Singer and John B. Willett
}
\usage{
data(reading)
}
\description{
The data were collected to investigate the change in reading scores over time
}
\examples{
data(reading)
}
\keyword{datasets}
