% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reeds.R
\docType{data}
\name{reeds}
\alias{reeds}
\title{Fowlmere reeds Data}
\format{A data frame with 2 columns and 14 rows. The columns are:
\itemize{
\item{site}{A (factor) label identifying the site.}
\item{nitrogen}{nitrogen content as a percentage.}}}
\usage{
data(reeds)
}
\description{
Percent nitrogen content of reeds sampled from three sites in Fowlmere, Cambridgeshire
}
\details{
Nitrogen content of reeds was investigated as part of a study of Moths in 1996.
}
\examples{
data(reeds)

}
\references{
Biomeasurement: A Student's Guide to Biological Statistics by Dawn Hawkins
}
\keyword{datasets}

