% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{spline.mixed}
\alias{spline.mixed}
\title{Mixed splines}
\usage{
spline.mixed(z, degree = 3, Nknots = NULL, type = c("TPB", "OSS"))
}
\arguments{
\item{z}{}

\item{degree}{}

\item{Nknots}{}

\item{type}{}
}
\description{
Mixed splines
}
