% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usair.R
\docType{data}
\name{usair}
\alias{usair}
\title{US Air Pollution}
\format{A data frame with 7 columns and 41 rows. The columns are:
\itemize{
\item{City}{Name of the US city}
\item{SO2}{Sulphur dioxide content of air in micrograms per cubic meter.}
\item{Neg.Temp}{Negative value of Average annual temperature in fahrenheit.}
\item{Manuf}{Number of manufacturing enterprises employing 20 or more workers.}
\item{Pop}{Population size (1970 census) in thousands.}
\item{Wind}{Average annual wind speed in miles per hour.}
\item{Precip}{Average annual precipitation in inches.}
\item{Days}{Average number of days with precipitation per year}
}}
\source{
Sokal and Rohlf (1981) Biometry: The Principles and Practices of Statistics in Biological Research
}
\usage{
data(usair)
}
\description{
The data were collected to investigate the determinants of pollution by considering SO2
level as the dependent variable and the remaining variables being potential explanatory variables.
}
\details{
Note that Neg.Temp represents the negative value of Average annual temperature.
}
\examples{
data(usair)
}
\references{
Hand, D. J., Daly, F., Lunn, A. D., McConway, K. J. and Ostrowski, E. (1994),
A handbook of small data sets, Chapman and Hall, London.

Brian S Everitt. An R and S-PLUS companion to multivariate analysis.
Springer Science & Business Media, 2006.
}
\keyword{datasets}
