% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer.R
\docType{data}
\name{wafer}
\alias{wafer}
\title{resitivity of wafer in semiconductor experiment}
\format{A data frame with 16 observations on the following 5 variables.
\describe{ \item{x1}{a factor with levels \code{-} \code{+}} \item{x2}{a
factor with levels \code{-} \code{+}} \item{x3}{a factor with levels
\code{-} \code{+}} \item{x4}{a factor with levels \code{-} \code{+}}
\item{resist}{Resistivity of the wafer} }}
\source{
Myers, R. and Montgomery D. (1997) A tutorial on generalized linear
models, Journal of Quality Technology, 29, 274-291.
}
\usage{
wafer
}
\description{
A full factorial experiment with four two-level predictors.
}
\keyword{datasets}
