% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XW.R
\name{BayesLM.nprior}
\alias{BayesLM.nprior}
\title{Bayesian Regression with Noninformative Priors}
\usage{
BayesLM.nprior(lmfit, B)
}
\arguments{
\item{lmfit}{lm output}

\item{B}{samples}
}
\value{
posteriors and summary
}
\description{
Bayesian Regression with Noninformative Priors
}
