% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr.R
\name{bri.band.ggplot}
\alias{bri.band.ggplot}
\title{Plot credible bands for a nonlinear function}
\usage{
bri.band.ggplot(result, name = NULL, alpha = 0.05, ind = NULL,
  type = c("random", "fitted", "linear"), xlab = NULL, ylab = NULL,
  main = NULL, hpd = FALSE, plot.data = FALSE)
}
\arguments{
\item{plot.data}{}
}
\description{
Plot credible bands for a nonlinear function
}

