% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XF.R
\name{bri.basehaz.plot}
\alias{bri.basehaz.plot}
\title{compute the baseline hazard function in INLA survival models}
\usage{
bri.basehaz.plot(inla.obj, plot = TRUE, ...)
}
\arguments{
\item{inla.obj}{}

\item{plot}{}

\item{...}{}
}
\description{
the function supports Weibull model, Expontenial model and Cox proportional hazards model for right censored data.
}

