% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XF.R
\name{bri.csresid.plot}
\alias{bri.csresid.plot}
\title{Cox-Snell residual plot}
\usage{
bri.csresid.plot(resid.obj, lwd = 2, xlab = "Cox-Snell Residual",
  ylab = "Estimated Cumulative Hazard Rates", main = "", ...)
}
\arguments{
\item{resid.obj}{}

\item{lwd}{}

\item{xlab}{}

\item{ylab}{}

\item{main}{}

\item{...}{}
}
\description{
Cox-Snell residual plot
}

