% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XW.R
\name{bri.dresid.plot}
\alias{bri.dresid.plot}
\title{Deviance residual plot}
\usage{
bri.dresid.plot(resid.obj, covariate = NULL, smooth = FALSE, xlab = NULL,
  ylab = "Deviance residual", main = "", cex.lab = 1.25,
  cex.axis = 1.25, ...)
}
\arguments{
\item{resid.obj}{object from bri.surv.resid function}

\item{covariate}{the covariate of interest for the residual plot.}

\item{smooth}{if smooth = TRUE, a nonparametric smooth curve will be added.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{main}{an overall title for the plot.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.axis}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{...}{}
}
\description{
Deviance residual plot
}
