% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{bri.excursions.ggplot}
\alias{bri.excursions.ggplot}
\title{Plot excursion sets and contour credible regions}
\usage{
bri.excursions.ggplot(res.exc, xlab = NULL, ylab = NULL, main = NULL)
}
\arguments{
\item{res.exc}{result object from excursions.brinla()}

\item{xlab}{a title for x axis}

\item{ylab}{a title for y axis}

\item{main}{an overall title for the plot}
}
\value{
a plot that shows the excursion set, the corresponding excursion function and the marginal excursion probabilities
}
\description{
Plot excursion sets and contour credible regions
}
