% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JJF.R
\name{bri.fixed.plot}
\alias{bri.fixed.plot}
\title{Plot posterior densities of the fixed effects}
\usage{
bri.fixed.plot(r, together = FALSE)
}
\arguments{
\item{r}{an inla model object}
}
\value{
a data frame containing the densities and parameter labels (invisible)
}
\description{
Plot posterior densities of the fixed effects
}
\author{
Julian Faraway, \email{jjf23@bath.ac.uk}
}
