% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XW.R
\name{bri.surv.resid}
\alias{bri.surv.resid}
\title{Compute different Bayesian residuals for survival models using INLA}
\usage{
bri.surv.resid(inla.obj, time, event)
}
\arguments{
\item{inla.obj}{an object of class "inla".}

\item{time}{the follow up time for right censored data,}

\item{event}{the status indicator, 1=observed event, 0=right censored event}
}
\value{
A list of different residuals, and time, event, family information.
}
\description{
Compute different Bayesian residuals for survival models using INLA
}
\details{
The function supports Weibull model, Expontenial model and Cox proportional hazards model for right censored data. Three types of residuals are obtained: Cox-Snell residual; Martingale residuals; Deviance residuals.
}
