% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XF.R
\name{bri.surv.resid}
\alias{bri.surv.resid}
\title{Residuals for inla survival models}
\usage{
bri.surv.resid(inla.obj, time, event)
}
\arguments{
\item{inla.obj}{}

\item{time}{}

\item{event}{}
}
\description{
the function supports Weibull model, Expontenial model and Cox proportional hazards model for right censored data.
Three types of residuals could be obtained: Cox-Snell residual; Martingale residuals; Deviance residuals.
}

