% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{bri.tps.prior}
\alias{bri.tps.prior}
\title{TPS Prior}
\usage{
bri.tps.prior(mesh, constr = FALSE, extraconstr.int = NULL,
  extraconstr = NULL, theta.mean = 0, theta.prec = 0.001)
}
\arguments{
\item{theta.prec}{}
}
\description{
TPS Prior
}
