% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{excursions.brinla}
\alias{excursions.brinla}
\title{Excursions}
\usage{
excursions.brinla(result.inla, name = NULL, ind = NULL, method, u, type,
  alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\description{
Excursions
}
