% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YY.R
\name{spline.mixed}
\alias{spline.mixed}
\title{Make matrices in the linear mixed models for truncated power basis splines and O'Sullivan splines}
\usage{
spline.mixed(z, degree = 3, Nknots = NULL, type = c("TPB", "OSS"))
}
\arguments{
\item{z}{the predictor vector}

\item{degree}{the degree of B-spline basis}

\item{Nknots}{the number of knots used in B-spline basis; the default number will be used if not specified}

\item{type}{the type of spline models: 'TPB' truncated power basis splines; 'OSS' O'Sullivan splines}
}
\value{
X: matrix for fixed effects; Z: matrix for random effects; Q: penalty matrix in O'Sullivan splines; B: matrix of B-spline basis functions
}
\description{
Make matrices in the linear mixed models for truncated power basis splines and O'Sullivan splines
}
